+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+- `disabled` property is now mapped to the TinyMCE `disabled` option.
+
+### Added
+- Added `readonly` property that maps to the TinyMCE `readonly` option.
+
 ## 6.1.0 - 2025-03-31
 
 ### Added